Enhancement: Now have a generalized error handler for errors.
