+++ song-server/src/main/java/bio/overture/song/server/config/ErrorConfig.java
@@ -1,0 +1,14 @@
+package bio.overture.song.server.config;
+
+import org.springframework.boot.autoconfigure.web.ErrorProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ErrorConfig {
+
+  @Bean
+  public ErrorProperties myErrorProperties() {
+    return new ErrorProperties();
+  }
+}
