+++ song-server/src/main/java/bio/overture/song/server/controller/StudyController.java
@@ -100,81 +100,5 @@ public class StudyController {
     return new ResponseEntity<>(format("{\"id\": \"%s\"}",id), HttpStatus.OK);
   }
 
-  @ExceptionHandler({ ServerException.class })
-  public ResponseEntity<Object> handleServerException(HttpServletRequest req,
-    ServerException ex) {
-    return errorResponse(HttpStatus.INTERNAL_SERVER_ERROR,
-      ex.getMessage(),
-      ex);
-  }
-  @ExceptionHandler({ Exception.class })
-  public ResponseEntity<Object> handleGeneralException(HttpServletRequest req,
-    Exception ex) {
-    return errorResponse(HttpStatus.INTERNAL_SERVER_ERROR,
-      "SONG INTERNAL ERROR",
-      ex);
-  }
 
-  @ExceptionHandler({ ConnectException.class })
-  public ResponseEntity<Object> handleConnectException(
-    HttpServletRequest req,
-    ResourceAccessException ex) {
-    return errorResponse(HttpStatus.INTERNAL_SERVER_ERROR,
-      "SONG can't connect to Ego", ex);
-  }
-
-  @ExceptionHandler({ ResourceAccessException.class })
-  public ResponseEntity<Object> handleResourceAccessException(HttpServletRequest req,
-    ResourceAccessException ex) {
-    return errorResponse(HttpStatus.INTERNAL_SERVER_ERROR,
-      "ResourceAccessException: %s", ex);
-  }
-
-  @ExceptionHandler({ HttpClientErrorException.class })
-  public ResponseEntity<Object> handleHttpException(HttpServletRequest req,
-    HttpClientErrorException ex) {
-    if (ex.getStatusCode() == HttpStatus.UNAUTHORIZED) {
-      return errorResponse(HttpStatus.UNAUTHORIZED,
-        "Not Authorized: %s", ex);
-    }
-
-    return errorResponse(ex.getStatusCode(),
-      "HTTP CLIENT Error: %s",
-      ex);
-  }
-
-  @ExceptionHandler({ InvalidTokenException.class })
-  public ResponseEntity<Object> handleTokenException(HttpServletRequest req,
-    InvalidTokenException ex) {
-    return errorResponse(HttpStatus.UNAUTHORIZED,
-      "Token Error: %s",
-      ex);
-  }
-
-  @ExceptionHandler({ AuthenticationException.class })
-  public ResponseEntity<Object> handleAuthenticationException(HttpServletRequest req,
-    InvalidTokenException ex) {
-    return errorResponse(HttpStatus.UNAUTHORIZED,
-      "Authentication Error: %s",
-      ex);
-  }
-
-  @ExceptionHandler({ EgoIsDownException.class })
-  public ResponseEntity<Object> handleSongDownException(
-    HttpServletRequest req,
-    EgoIsDownException ex) {
-    return errorResponse(HttpStatus.INTERNAL_SERVER_ERROR,
-      "Ego is down: %s",
-      ex);
-  }
-
-  private String jsonEscape(String text) {
-    return text.replace("\"", "\\\"");
-  }
-
-  private ResponseEntity<Object> errorResponse(HttpStatus status, String msg, Exception ex) {
-    log.error(format("Creating error response for exception: '%s'", ex.getMessage()));
-    val json = format("{\"error_message\": \"%s\"}", msg);
-    return new ResponseEntity<>(json, status);
-  }
 }
